+++ module/model-sql/__index__.ts
@@ -3,4 +3,5 @@ export * from './src/connection.ts';
 export * from './src/dialect.ts';
 export * from './src/query.ts';
 export * from './src/service.ts';
+export * from './src/types.ts';
 export * from './src/util.ts';
